Utilities cleanup.
