+++ backend/main/serializers/clientstate/ser_client_state.py
@@ -2,7 +2,7 @@ from rest_framework import serializers
 from main.models import User, Board
 from main.validation.val_auth import authentication_error, authorization_error
 from main.validation.val_custom import CustomAPIException
-from main.util import create_board
+from main.utilities import create_board
 import bcrypt
 import status
