+++ backend/main/tests/test_post_users.py
@@ -1,7 +1,7 @@
 from rest_framework.test import APITestCase
 from rest_framework.exceptions import ErrorDetail
 from ..models import User, Board, Team
-from ..util import create_admin
+from ..utilities import create_admin
 from ..validation.val_auth import authentication_error, authorization_error
